+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 - (Feature) Add new field to DeploymentReplicationStatus with details on DC2DC sync status
 - (Feature) Early connections support
 - (Bugfix) Fix and document action timeouts
-
 - (Feature) Propagate sidecars' ports to a member's service
+- (Debug Package) Initial commit
 
 ## [1.2.16](https://github.com/arangodb/kube-arangodb/tree/1.2.16) (2022-09-14)
 - (Feature) Add ArangoDeployment ServerGroupStatus